backend/requirements.txt
Django>=5.1,<6.0
djangorestframework>=3.16.1
channels>=4.1
channels_redis>=4.0
psycopg2-binary>=2.9
django-environ>=0.10
django-cors-headers>=4.0
fastapi>=0.111
uvicorn[standard]>=0.23
httpx
openai>=1.0
requests>=2.31
djangorestframework-simplejwt>=2.2
python-dotenv>=1.0
pydantic>=2.3
pytest>=8.0
pytest-django>=4.5
black>=24.10
isort>=6.0
tavily
PyPDF2
python-docx
python-pptx